refactor(Budget): money를 long에서 int로 변경

+++ src/main/java/lotto/model/Budget.java
@@ -5,20 +5,20 @@ import lotto.utils.ExceptionMessage;
 public class Budget {
     private static final int BASE_MONEY = 1000;
 
-    private final long money;
+    private final int money;
 
-    public Budget(long money) {
+    public Budget(int money) {
         validate(money);
         this.money = money;
     }
 
-    private void validate(long money) {
+    private void validate(int money) {
         if (money % BASE_MONEY != 0) {
             throw new IllegalArgumentException(ExceptionMessage.INVALID_INPUT_BUDGET.getMessage());
         }
     }
 
-    public long calculateNumberOfLotto() {
+    public int calculateNumberOfLotto() {
         return money / BASE_MONEY;
     }
 }
